+++ packages/vfx/src/layers/animateScale.ts
@@ -1,4 +1,4 @@
-import { createShader } from "../newShaders"
+import { createShader } from "../lib/shadermaker"
 
 export default function(fun = "v_progress") {
   return createShader({
