+++ packages/vfx/src/layers/provideResolution.ts
@@ -1,4 +1,4 @@
-import { createShader, Uniform } from "../newShaders"
+import { createShader, Uniform } from "../lib/shadermaker"
 
 export default function() {
   const u_resolution: Uniform = {
